RecoveredProject-0/src/test/java/services/AccountServicesTest.java
package services;

import dao.AccountsDAO;
import models.Account;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;

import java.util.ArrayList;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class AccountServicesTest {

    AccountsDAO accountDao = Mockito.mock(AccountsDAO.class);

    AccountServices accountServices;

    public AccountServicesTest(){this.accountServices = new AccountServices(accountDao);}

    @Test
    void createAccount() {
        //arrange
        Account account = new Account(1, 10, 23.3);

        //act
        accountServices.createAccount(account);

        //assert
        Mockito.verify(accountDao, Mockito.times(1)).createAccount(account);
    }

    @Test
    void getAllAccounts() {
        //arrange
        List<Account> accountList = new ArrayList<>();
        accountList.add(new Account(12, 10, 200d));
        accountList.add(new Account(13, 10, 600d));
        //accountList.add(new Account(3, 3, 200d));
        accountList.add(new Account(14, 10, 232.21));
        //accountList.add(new Account(5, 1, 70d));
        //accountList.add(new Account(6, 1, 2230d));
        List<Account> expectedValue = accountList;
        Mockito.when(accountServices.getAllAccounts(10)).thenReturn(accountList);

        //act
        List<Account> actualResult = accountServices.getAllAccounts(10);

        //assert
        assertEquals(expectedValue,actualResult);
    }

    @Test
    void getSpecificAccounts() {
        //arrange
        List<Account> accountList = new ArrayList<>();
        accountList.add(new Account(12, 10, 200d));
        //accountList.add(new Account(13, 10, 600d));
        //accountList.add(new Account(3, 3, 200d));
        accountList.add(new Account(14, 10, 232.21));
        //accountList.add(new Account(5, 1, 70d));
        //accountList.add(new Account(6, 1, 2230d));
        List<Account> expectedValue = accountList;
        Mockito.when(accountServices.getSpecificAccounts(10,500d,100d)).thenReturn(accountList);

        //act
        List<Account> actualResult = accountServices.getSpecificAccounts(10,500d,100d);

        //assert
        assertEquals(expectedValue,actualResult);
    }

    @Test
    void getOneAccount() {
        //arrange
        /*
        List<Account> accountList = new ArrayList<>();
        accountList.add(new Account(14, 10, 232.21));
        //accountList.add(new Account(12, 10, 200d));
        //accountList.add(new Account(13, 10, 600d));
        //accountList.add(new Account(3, 3, 200d));
        //accountList.add(new Account(5, 1, 70d));
        //accountList.add(new Account(6, 1, 2230d));
        List<Account> expectedValue = accountList;
        Mockito.when(accountServices.getOneAccount(10)).thenReturn();

        //act
        List<Account> actualResult = accountServices.getSpecificAccounts(10,500d,100d);

        //assert
        assertEquals(expectedValue,actualResult); */
    }

    @Test
    void updateAccount() {//does not work
        //arrange
        //List<Account> accountList = new ArrayList<>();
        //accountList.add(new Account(12, 10, 20d));
        //accountList.add(new Account(13, 10, 600d));
        //accountList.add(new Account(3, 3, 200d));
        //accountList.add(new Account(14, 10, 232.21));
        //accountList.add(new Account(5, 1, 70d));
        //accountList.add(new Account(6, 1, 2230d));
        Integer accountId = 12;

        //act
        //accountServices.updateAccount(accountId,20d,-1.3); //hits a null pointer exception because of the scope

        //assert
        //Mockito.verify(accountDao, Mockito.times(1)).updateAccount(12, 20d);
    }

    @Test
    void deleteAccount() {
        //arrange
        Integer accountId = 12;

        //act
        accountServices.deleteAccount(accountId);

        //assert
        Mockito.verify(accountDao, Mockito.times(1)).deleteAccount(12);
    }

    @Test
    void transferMoney() {
        //same issue with verify, it is because of the getbal() method in transfer money
    }
}